Conflicts:
	carton.lock
	conf/crontab.ugly
	conf/general.yml-example
	conf/packages
	db/schema.sql
	perllib/FixMyStreet/App/Controller/Admin.pm
	perllib/FixMyStreet/App/Controller/Report/New.pm
	perllib/FixMyStreet/DB/Result/Open311conf.pm
	templates/web/fixmystreet/report/new/councils_text_all.html
	web/fixmystreet_app_cgi.cgi
	web/fixmystreet_app_fastcgi.cgi

+++ t/app/controller/alert_new.t
@@ -142,25 +142,18 @@ foreach my $test (
 }
 
 foreach my $test (
-    {
-        email      => 'test-new@example.com',
-        type       => 'area',
-        content    => 'your alert will not be activated',
-        email_text => 'confirm the alert',
-        uri =>
-'/alert/subscribe?type=local&rznvy=test-new@example.com&feed=area:1000:A_Location',
-        param1 => 1000
-    }
+    { exist => 0 },
+    { exist => 1 },
   )
 {
-    subtest "use existing unlogged in user in a alert" => sub {
+    subtest "use existing unlogged in user in a alert ($test->{exist})" => sub {
         $mech->log_out_ok();
 
-        my $type = $test->{type} . '_problems';
+        my $type = 'area_problems';
 
         my $user =
           FixMyStreet::App->model('DB::User')
-          ->find_or_create( { email => $test->{email} } );
+          ->find_or_create( { email => 'test-new@example.com' } );
 
         my $alert = FixMyStreet::App->model('DB::Alert')->find(
             {
@@ -169,24 +162,26 @@ foreach my $test (
             }
         );
         # clear existing data so we can be sure we're creating it
-        ok $alert->delete() if $alert;
+        ok $alert->delete() if $alert && !$test->{exist};
 
-        $mech->get_ok( $test->{uri} );
+        $mech->get_ok( '/alert/subscribe?type=local&rznvy=test-new@example.com&feed=area:1000:A_Location' );
 
         $alert = FixMyStreet::App->model('DB::Alert')->find(
             {
                 user       => $user,
                 alert_type => $type,
-                parameter  => $test->{param1},
-                parameter2 => $test->{param2},
-                confirmed  => 0,
+                parameter  => 1000,
+                parameter2 => undef,
+                confirmed  => $test->{exist},
             }
         );
 
         $mech->content_contains( 'Now check your email' );
 
+        $alert->confirm();
         ok $alert, 'New alert created with existing user';
-        $mech->delete_user($user);
+
+        $mech->delete_user($user) if $test->{exist};
     };
 }
 
@@ -445,6 +440,7 @@ subtest "Test normal alert signups and that alerts are sent" => sub {
     ok $update, "created test update - $update_id";
 
     FixMyStreet::App->model('DB::AlertType')->email_alerts();
+    # TODO Note the below will fail if the db has an existing alert that matches
     $mech->email_count_is(3);
     my @emails = $mech->get_email;
     my $count;
@@ -452,7 +448,7 @@ subtest "Test normal alert signups and that alerts are sent" => sub {
         $count++ if $_->body =~ /The following updates have been left on this problem:/;
         $count++ if $_->body =~ /The following new problems have been reported to City of\s*Edinburgh Council:/;
         $count++ if $_->body =~ /The following nearby problems have been added:/;
-        $count++ if $_->body =~ / -\s+Testing, EH1 1BB/;
+        $count++ if $_->body =~ /\s+-\s+Testing/;
     }
     is $count, 5, 'Five emails with the right things in them';
 
@@ -466,6 +462,196 @@ subtest "Test normal alert signups and that alerts are sent" => sub {
 
     $mech->delete_user($user1);
     $mech->delete_user($user2);
+};
+
+for my $test (
+    {
+        desc => 'check non public reports are not included in council problems alerts',
+        alert_params => {
+            alert_type => 'council_problems',
+            parameter => '2651',
+            parameter2 => '2651',
+        }
+    },
+    {
+        desc => 'check non public reports are not included in ward problems alerts',
+        alert_params => {
+            alert_type => 'ward_problems',
+            parameter => '2651',
+            parameter2 => '20728',
+        }
+    },
+    {
+        desc => 'check non public reports are not included in local problems alerts',
+        alert_params => {
+            alert_type => 'local_problems',
+            parameter => '-3.189944',
+            parameter2 => '55.951963',
+        }
+    },
+    {
+        desc => 'check non public reports are not included in area problems alerts',
+        alert_params => {
+            alert_type => 'area_problems',
+            parameter => '20728',
+            parameter2 => '20728',
+        }
+    },
+) {
+    subtest $test->{desc} => sub {
+        my $user1 = FixMyStreet::App->model('DB::User')
+          ->find_or_create( { email => 'reporter@example.com', name => 'Reporter User' } );
+        ok $user1, "created test user";
+        $user1->alerts->delete;
+
+        my $user2 = FixMyStreet::App->model('DB::User')
+          ->find_or_create( { email => 'alerts@example.com', name => 'Alert User' } );
+        ok $user2, "created test user";
+        $user2->alerts->delete;
+
+        my $dt = DateTime->now->add( minutes => -30 );
+        my $r_dt = $dt->clone->add( minutes => 20 );
+
+        my $alert_params = $test->{alert_params};
+        $alert_params->{user} = $user1;
+        $alert_params->{whensubscribed} = $dt;
+        $alert_params->{confirmed} = 1;
+
+        my $alert_user1 = FixMyStreet::App->model('DB::Alert')->create( $alert_params );
+        ok $alert_user1, "alert created";
+
+        my $report = FixMyStreet::App->model('DB::Problem')->find_or_create( {
+            postcode           => 'EH1 1BB',
+            council            => '2651',
+            areas              => ',11808,135007,14419,134935,2651,20728,',
+            category           => 'Street lighting',
+            title              => 'Alert test for non public reports',
+            detail             => 'Testing Detail',
+            used_map           => 1,
+            name               => $user2->name,
+            anonymous          => 0,
+            state              => 'confirmed',
+            confirmed          => $r_dt,
+            lastupdate         => $r_dt,
+            whensent           => $r_dt->clone->add( minutes => 5 ),
+            lang               => 'en-gb',
+            service            => '',
+            cobrand            => 'default',
+            cobrand_data       => '',
+            send_questionnaire => 1,
+            latitude           => '55.951963',
+            longitude          => '-3.189944',
+            user_id            => $user2->id,
+            non_public         => 1,
+        } );
+
+        $mech->clear_emails_ok;
+        FixMyStreet::App->model('DB::AlertType')->email_alerts();
+        $mech->email_count_is(0);
+
+        $report->update( { non_public => 0 } );
+        FixMyStreet::App->model('DB::AlertType')->email_alerts();
+        $mech->email_count_is(1);
+        my $email = $mech->get_email;
+        like $email->body, qr/Alert\s+test\s+for\s+non\s+public\s+reports/, 'alert contains public report';
+
+        $mech->delete_user( $user1 );
+        $mech->delete_user( $user2 );
+    };
+}
+
+subtest 'check new updates alerts for non public reports only go to report owner' => sub {
+    my $user1 = FixMyStreet::App->model('DB::User')
+      ->find_or_create( { email => 'reporter@example.com', name => 'Reporter User' } );
+    ok $user1, "created test user";
+    $user1->alerts->delete;
+
+    my $user2 = FixMyStreet::App->model('DB::User')
+      ->find_or_create( { email => 'alerts@example.com', name => 'Alert User' } );
+    ok $user2, "created test user";
+    $user2->alerts->delete;
+
+    my $user3 = FixMyStreet::App->model('DB::User')
+      ->find_or_create( { email => 'updates@example.com', name => 'Update User' } );
+    ok $user3, "created test user";
+
+    my $dt = DateTime->now->add( minutes => -30 );
+    my $r_dt = $dt->clone->add( minutes => 20 );
+
+    my $report = FixMyStreet::App->model('DB::Problem')->find_or_create( {
+        postcode           => 'EH1 1BB',
+        council            => '2651',
+        areas              => ',11808,135007,14419,134935,2651,20728,',
+        category           => 'Street lighting',
+        title              => 'Alert test for non public reports',
+        detail             => 'Testing Detail',
+        used_map           => 1,
+        name               => $user2->name,
+        anonymous          => 0,
+        state              => 'confirmed',
+        confirmed          => $r_dt,
+        lastupdate         => $r_dt,
+        whensent           => $r_dt->clone->add( minutes => 5 ),
+        lang               => 'en-gb',
+        service            => '',
+        cobrand            => 'default',
+        cobrand_data       => '',
+        send_questionnaire => 1,
+        latitude           => '55.951963',
+        longitude          => '-3.189944',
+        user_id            => $user2->id,
+        non_public         => 1,
+    } );
+
+    my $update = FixMyStreet::App->model('DB::Comment')->create( {
+        problem_id => $report->id,
+        user_id    => $user3->id,
+        name       => 'Anonymous User',
+        mark_fixed => 'false',
+        text       => 'This is some more update text',
+        state      => 'confirmed',
+        confirmed  => $r_dt->clone->add( minutes => 8 ),
+        anonymous  => 't',
+    } );
+
+    my $alert_user1 = FixMyStreet::App->model('DB::Alert')->create( {
+            user       => $user1,
+            alert_type => 'new_updates',
+            parameter  => $report->id,
+            confirmed  => 1,
+            whensubscribed => $dt,
+    } );
+    ok $alert_user1, "alert created";
+
+
+    $mech->clear_emails_ok;
+    FixMyStreet::App->model('DB::AlertType')->email_alerts();
+    $mech->email_count_is(0);
+
+    my $alert_user2 = FixMyStreet::App->model('DB::Alert')->create( {
+            user       => $user2,
+            alert_type => 'new_updates',
+            parameter  => $report->id,
+            confirmed  => 1,
+            whensubscribed => $dt,
+    } );
+    ok $alert_user2, "alert created";
+
+    FixMyStreet::App->model('DB::AlertType')->email_alerts();
+    $mech->email_count_is(1);
+    my $email = $mech->get_email;
+    like $email->body, qr/This is some more update text/, 'alert contains update text';
+
+    $mech->clear_emails_ok;
+    $report->update( { non_public => 0 } );
+    FixMyStreet::App->model('DB::AlertType')->email_alerts();
+    $mech->email_count_is(1);
+    $email = $mech->get_email;
+    like $email->body, qr/This is some more update text/, 'alert contains update text';
+
+    $mech->delete_user( $user1 );
+    $mech->delete_user( $user2 );
+    $mech->delete_user( $user3 );
 };
 
 done_testing();
